--- app/src/main/java/com/babylon/wallet/android/data/transaction/TransactionConfig.kt
@@ -1,7 +1,0 @@
-package com.babylon.wallet.android.data.transaction
-
-object TransactionConfig {
-    const val TIP_PERCENTAGE: UShort = 0u
-    const val EPOCH_WINDOW = 10u
-    const val DEFAULT_LOCK_FEE = 25L
-}
