+++ app/src/main/java/com/babylon/wallet/android/domain/usecases/transaction/TransactionConfig.kt
@@ -1,0 +1,8 @@
+package com.babylon.wallet.android.domain.usecases.transaction
+
+// TODO move them to domain since they are used only in usecases and viewmodels
+object TransactionConfig {
+    const val TIP_PERCENTAGE: UShort = 0u
+    const val EPOCH_WINDOW = 10u
+    const val DEFAULT_LOCK_FEE = 25L
+}
